snapshot: - move ngal to base galaxy catalog - refactor positions - move transforms to dedicated file - move tests at the correct place - generalize _nbar function - Make downsample use volume instead of boxsize
