+++ acm/catalogs/products/snapshot.py
@@ -10,135 +10,9 @@ from pandas._typing import RandomState
 
 from acm.catalogs.dataclasses import Transform
 from acm.catalogs.products import BaseGalaxyCatalog
+from acm.catalogs.products.transforms import _apply_ap, _apply_rsd, _apply_downsample
 
 logger = logging.getLogger(__name__)
-
-
-# %% Pure transform functions that can be used in the pipeline
-def _apply_rsd(data: pd.DataFrame, los: str, hubble: float, az: float) -> pd.DataFrame:
-    """
-    Apply RSD shift along the los axis.
-
-    Parameters
-    ----------
-    data : pd.DataFrame
-        Galaxy data containing position and velocity columns.
-    los : str
-        Line-of-sight axis, one of 'x', 'y', 'z'.
-    hubble : float
-        Hubble parameter H(z) in km/s/(Mpc/h) for the simulation cosmology.
-    az : float
-        Scale factor a(z) at the snapshot's redshift.
-
-    Returns
-    -------
-    pd.DataFrame
-        Transformed galaxy data with RSD applied.
-    """
-    data = data.copy()
-    v_col = f"v{los}"
-    data[los] = data[los] + data[v_col] / (hubble * az)
-    return data
-
-
-def _apply_ap(
-    data: pd.DataFrame,
-    los: str,
-    q_par: float,
-    q_perp: float,
-    pos_columns: tuple[str],
-) -> pd.DataFrame:
-    """
-    Apply AP scaling: q_par along los, q_perp along transverse axes.
-
-    Parameters
-    ----------
-    data : pd.DataFrame
-        Galaxy data containing position columns.
-    los : str
-        Line-of-sight axis, one of 'x', 'y', 'z'.
-    q_par : float
-        AP scaling factor along the line-of-sight.
-    q_perp : float
-        AP scaling factor along the transverse directions.
-    pos_columns : tuple[str]
-        Names of the position columns, e.g. ('x', 'y', 'z').
-
-    Returns
-    -------
-    pd.DataFrame
-        Transformed galaxy data with AP scaling applied.
-    """
-    data = data.copy()
-    for ax in pos_columns:
-        data[ax] = data[ax] * (q_par if ax == los else q_perp)
-    return data
-
-
-def _apply_downsample(
-    data: pd.DataFrame,
-    tracer: str,
-    n_gal: int | None,
-    f_gal: float | None,
-    nbar: float | None,
-    boxsize: Callable[[], np.ndarray] | None = None,
-    seed: RandomState | None = None,
-) -> pd.DataFrame:
-    """
-    Randomly downsample a tracer DataFrame.
-
-    Parameters
-    ----------
-    data : pd.DataFrame
-        Galaxy data for a specific tracer.
-    tracer : str
-        Tracer name, used for logging.
-    n_gal : int, optional
-        Target number of galaxies.
-    f_gal : float, optional
-        Fraction of galaxies to keep, between 0 and 1.
-    nbar : float, optional
-        Target number density in (Mpc/h)^-3.
-    boxsize : callable, optional
-        Function that returns the current boxsize, needed to compute target n_gal when downsampling by nbar.
-
-    Returns
-    -------
-    pd.DataFrame
-        Downsampled galaxy data.
-
-    Raises
-    ------
-    ValueError
-        If not exactly one of n_gal, f_gal or nbar is provided.
-    ValueError
-        If boxsize is needed but not provided.
-    """
-    provided = sum(p is not None for p in (n_gal, f_gal, nbar))
-    if provided != 1:
-        raise ValueError("Exactly one of n_gal,f_gal or nbar must be provided.")
-
-    n_current = len(data)
-    if f_gal is not None:
-        n_target = round(n_current * f_gal)
-    elif n_gal is not None:
-        n_target = n_gal
-    else:  # nbar
-        if boxsize is None:
-            raise ValueError(
-                "boxsize function must be provided when downsampling by nbar."
-            )
-        # Callable to get current boxsize, which may include AP scaling
-        volume = np.prod(boxsize())
-        n_target = round(nbar * volume)
-
-    if n_target >= n_current:
-        logger.warning(
-            f"Target n_gal={n_target} >= current n_gal={n_current} for tracer '{tracer}', skipping downsample."
-        )
-        return data
-
-    return data.sample(n=n_target, random_state=seed).reset_index(drop=True)
 
 
 # %% GalaxyCatalog classes
@@ -339,25 +213,13 @@ class SnapshotCatalog(BaseGalaxyCatalog):
                     "n_gal": n_gal,
                     "f_gal": f_gal,
                     "nbar": nbar,
-                    "boxsize": lambda: self.boxsize,  # evaluated at application time
+                    "volume": lambda: np.prod(self.boxsize),  # evaluated at application time
                     "seed": seed,
                 },
             )
         )
 
-    def _ngal(self, tracer: str) -> int:
-        """Return the total number of galaxies for a specific tracer."""
-        d = self.get_tracer_data(tracer)
-        return len(d)
-
-    @property
-    def ngal(self) -> int:
-        """Total number of galaxies in the catalog across all tracers."""
-        if not self.tracers:
-            raise RuntimeError("No tracers loaded in the catalog, cannot compute ngal.")
-        return sum(self._ngal(tracer) for tracer in self.tracers)
-
-    def _nbar(self, tracer: str) -> float:
+    def _nbar(self, tracer: str | None = None) -> float:
         """Return the number density of galaxies for a specific tracer."""
         n_gal = self._ngal(tracer)
         boxsize = self.boxsize
@@ -367,10 +229,7 @@ class SnapshotCatalog(BaseGalaxyCatalog):
     @property
     def nbar(self) -> float:
         """Number density of galaxies in the entire catalog."""
-        if not self.tracers:
-            raise RuntimeError("No tracers loaded in the catalog, cannot compute nbar.")
-        volume = np.prod(self.boxsize)
-        return self.ngal / volume if volume > 0 else 0.0
+        return self._nbar()
 
     def positions(self, raw: bool = True) -> pd.DataFrame:
         """
@@ -391,13 +250,12 @@ class SnapshotCatalog(BaseGalaxyCatalog):
             raise RuntimeError(
                 "No tracers loaded in the catalog, cannot get positions."
             )
-        pos = []
-        for tracer in self.tracers:
-            d = self.get_tracer_data(tracer, raw=raw)
-            pos.append(d[list(self.pos_columns)])
-        return pd.concat(pos, ignore_index=True)
-
-    # TODO: Add box replocation with padding for cutsky creation ?
+        data = pd.concat(
+            [self.get_tracer_data(t, raw=raw) for t in self.tracers], 
+            ignore_index=True,
+        )
+        pos = data[list(self.pos_columns)]
+        return pos
 
     def _save_attrs(self, f: h5py.File) -> None:
         f.attrs["redshift"] = self.redshift
